# specs/001-ticket-dedup-dashboard/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: DedupTickets API
  description: API for the Ticket Deduplication & Clustering Dashboard
  version: 1.0.0
  contact:
    name: DedupTickets Team
servers:
  - url: /api/v1
    description: API v1
security:
  - ApiKeyAuth: []
...
paths:
...
  /spikes:
    get:
      tags: [Spikes]
      summary: List spike alerts
      operationId: listSpikes
      parameters:
        - $ref: '#/components/parameters/PageOffset'
        - $ref: '#/components/parameters/PageLimit'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/SpikeStatus'
        - name: severity
          in: query
          schema:
            $ref: '#/components/schemas/SeverityLevel'
        - name: field_name
          in: query
          schema:
            type: string
        - name: detected_after
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Paginated list of spike alerts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpikeListResponse'
  /spikes/{spike_id}:
    get:
      tags: [Spikes]
      summary: Get spike alert details
      operationId: getSpike
      parameters:
        - $ref: '#/components/parameters/SpikeId'
      responses:
        '200':
          description: Spike alert with related clusters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpikeDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /spikes/{spike_id}/acknowledge:
    post:
      tags: [Spikes]
      summary: Acknowledge a spike alert
      operationId: acknowledgeSpike
      parameters:
        - $ref: '#/components/parameters/SpikeId'
      responses:
        '200':
          description: Spike acknowledged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpikeAlert'
        '404':
          $ref: '#/components/responses/NotFound'
  /spikes/{spike_id}/resolve:
    post:
      tags: [Spikes]
      summary: Resolve a spike alert
      operationId: resolveSpike
      parameters:
        - $ref: '#/components/parameters/SpikeId'
      responses:
        '200':
          description: Spike resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpikeAlert'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== TRENDS ====================
...
  /audit:
    get:
      tags: [Audit]
      summary: Search audit logs
      operationId: searchAuditLogs
      parameters:
        - $ref: '#/components/parameters/PageOffset'
        - $ref: '#/components/parameters/PageLimit'
        - name: action_type
          in: query
          schema:
            $ref: '#/components/schemas/AuditActionType'
        - name: actor_id
          in: query
          schema:
            type: string
        - name: resource_type
          in: query
          schema:
            type: string
        - name: resource_id
          in: query
          schema:
            type: string
            format: uuid
        - name: created_after
          in: query
          schema:
            type: string
            format: date-time
        - name: created_before
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Paginated audit log entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditListResponse'
  /audit/{entry_id}:
    get:
      tags: [Audit]
      summary: Get audit entry details
      operationId: getAuditEntry
      parameters:
        - name: entry_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Audit entry details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEntry'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    PageOffset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    PageLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    SpikeId:
      name: spike_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
    PaginationMeta:
      type: object
      required: [total, offset, limit]
      properties:
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
        has_more:
          type: boolean

    # ==================== ENUMS ====================
...
    ClusterStatus:
      type: string
      enum: [pending, merged, dismissed, expired]
    ConfidenceLevel:
      type: string
      enum: [high, medium, low]
...
    SpikeStatus:
      type: string
      enum: [active, acknowledged, resolved]
    SeverityLevel:
      type: string
      enum: [low, medium, high]
    AuditActionType:
      type: string
      enum: [merge, revert, cluster_create, cluster_dismiss, cluster_member_remove, ticket_create, ticket_update, spike_detect, spike_acknowledge, spike_resolve]

    # ==================== TICKETS ====================
...
    MatchingSignals:
      type: object
      properties:
        exact_matches:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              value:
                type: string
        time_window:
          type: object
          properties:
            start:
              type: string
              format: date-time
            end:
              type: string
              format: date-time
        text_similarity:
          type: object
          properties:
            score:
              type: number
              format: float
            common_terms:
              type: array
              items:
                type: string
        field_matches:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              value:
                type: string
    Cluster:
      type: object
      required: [id, status, confidence, summary, ticket_count, created_at]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/ClusterStatus'
        confidence:
          $ref: '#/components/schemas/ConfidenceLevel'
        summary:
          type: string
        matching_signals:
          $ref: '#/components/schemas/MatchingSignals'
        primary_ticket_id:
          type: string
          format: uuid
        ticket_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
          description: Auto-expire if not actioned
        created_by:
          type: string
          description: '"system" or user ID'
...
    SpikeAlert:
      type: object
      required: [id, status, severity, field_name, field_value, current_count, baseline_count, percentage_increase, detected_at]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/SpikeStatus'
        severity:
          $ref: '#/components/schemas/SeverityLevel'
        field_name:
          type: string
        field_value:
          type: string
        current_count:
          type: integer
        baseline_count:
          type: number
        percentage_increase:
          type: number
        time_window_start:
          type: string
          format: date-time
        time_window_end:
          type: string
          format: date-time
        detected_at:
          type: string
          format: date-time
        acknowledged_by:
          type: string
        acknowledged_at:
          type: string
          format: date-time
        resolved_at:
          type: string
          format: date-time
    SpikeDetail:
      allOf:
        - $ref: '#/components/schemas/SpikeAlert'
        - type: object
          properties:
            affected_clusters:
              type: array
              items:
                $ref: '#/components/schemas/Cluster'
    SpikeListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/SpikeAlert'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    # ==================== TRENDS ====================
...
    AuditEntry:
      type: object
      required: [id, action_type, actor_id, actor_type, resource_type, resource_id, outcome, created_at]
      properties:
        id:
          type: string
          format: uuid
        action_type:
          $ref: '#/components/schemas/AuditActionType'
        actor_id:
          type: string
        actor_type:
          type: string
          enum: [user, system]
        resource_type:
          type: string
        resource_id:
          type: string
          format: uuid
        related_ids:
          type: array
          items:
            type: string
            format: uuid
        metadata:
          type: object
        outcome:
          type: string
          enum: [success, failure]
        error_message:
          type: string
        ip_address:
          type: string
          description: Client IP address
        user_agent:
          type: string
          description: Client user agent
        created_at:
          type: string
          format: date-time
    AuditListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AuditEntry'
        meta:
          $ref: '#/components/schemas/PaginationMeta'